feat(TDP-3333): refacto playground service
